// src/app/models/personal.model.ts
export class PersonalModel {
    // tslint:disable-next-line: variable-name
    id_usuario?: string;
    nombre: string;
    apellidos: string;
    telefono: string;
    correo: string;
    contrasena: string;
    // tslint:disable-next-line: variable-name
    rol_usuario: string;
    rfc: string;
    ocupacion: string;
    departamento: string;
}