// src/app/models/externo.model.ts
export class ExternoModel {
    // tslint:disable-next-line: variable-name
    id_usuario?: string;
    nombre: string;
    apellidos: string;
    telefono: string;
    correo: string;
    contrasena: string;
    // tslint:disable-next-line: variable-name
    rol_usuario: string;
    direccion: string;
}